minor tweaks, removed some warnings
git-svn-id: http://svn.xiph.org/trunk/speex@4510 0101bb08-14d6-0310-b084-bc0e0c8e3800

+++ libspeex/sb_celp.c
@@ -808,7 +808,7 @@ int sb_decode(void *state, SpeexBits *bits, float *out)
    char *stack;
    float *low_pi_gain, *low_exc, *low_innov;
    float *awk1, *awk2, *awk3;
-   float dtx;
+   int dtx;
    SpeexSBMode *mode;
 
    st = (SBDecState*)state;
